Move dialogue validation to the loader
